refactor(p2-5)!: 旧 api.ts 3 ファイル削除と time-entries cleanup
Gateway 層への移行完了に伴い、以下をクリーンアップする。

- entities/{task,project,event}/api.ts を削除（Gateway 経由で置換済み）
- entities/task/time-entries.ts から DB アクセス関数
  （listTimeEntries / getOpenTimeEntry / startTimeEntry / closeTimeEntry）
  と内部ヘルパーを削除し、純粋関数 sumDurationSeconds + 型
  （PauseReason / TimeEntry）のみに縮約
- features/add-forms/{AddPanel,ProjectForm,TaskForm,EventForm}.tsx の
  Input 型 import 先を @/entities/*/api → @/entities/*/gateway に移行

これにより src/ 配下の createClient 呼び出しは以下のみに限定される:
- auth flow (page.tsx / login / callback / UserMenu の signOut)
- Supabase factory (shared/supabase/{client,server}.ts)
- 新 Gateway factory (shared/gateway/createSupabaseGateways.ts)
- action-log logger singleton（設計 Note 明示のスコープ外、#47 で再評価）

挙動不変（cache key / optimistic / invalidate / TASK_DELETED 発火タイミング
すべて無変更）、116 tests 緑を維持。

BREAKING: entities/*/api.ts を削除した。外部から import していた場合は
@/entities/*/gateway（Input 型）および Gateway interface / SupabaseXxxGateway
経由に切り替える必要がある。

Closes #45
